requirements.txt
chess==1.10.0
selenium==4.22.0
pillow==10.4.0
opencv-python==4.10.0.84
pyautogui==0.9.54
numpy==2.0.0